# control/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "main.githubAppDTO": {
                "properties": {
                    "app_id": {
                        "type": "integer"
                    },
                    "callback_url": {
                        "type": "string"
                    },
                    "key_set": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "The self-service API: sign in, manage your profile and your personal access tokens, and create and run VMs.\n\nTwo credentials reach these routes. A browser sends the httpOnly `access_token` cookie. A script sends a personal access token as `Authorization: Bearer dpat_...` — mint one under Settings, or with `POST /me/tokens` from a signed-in session.\n\nA personal access token does everything you can do with your own VMs. It can never reach `/api/v1/admin/*`, even when it belongs to an admin, and it cannot manage tokens: those routes take a password-backed session, so a leaked token cannot mint its successor.\n\nAdmin and client-enrollment routes are deliberately absent here. The first is browser-only by design; the second is a protocol between the control plane and its hosts rather than an API.",
        "title": "dummie control API",
        "version": "1.0"
    },
...
    "paths": {
        "/admin/github-app": {
            "delete": {
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Forget the github app",
                "tags": [
                    "admin"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.githubAppDTO"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "The github app backing integrations",
                "tags": [
                    "admin"
                ]
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.githubAppDTO"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Configure the github app",
                "tags": [
                    "admin"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}